SpiralIterator fix for center out of bounds
In case the center of the SprialIterator was chosen to be outside of the map
a segmentation fault was thrown when trying to use the iterator.

One reason was, that the index of the center was not computed correctly,
because the getIndexFromPosition returned false without changing
the value for the index. Now the index is computed, but the function still
returns false in case the position is not inside the map.

The other reason for this was that only the cells next to the center were
looked at in that case. When these cells also were outside of the map, then
dereferencing the pointer lead to a segmentation fault. This was fixed by
generating rings until a position that is inside the map was reached.